+++ declavac/Cargo.toml
@@ -10,6 +10,9 @@ repository.workspace = true
 crate-type = ["cdylib"]
 
 [dependencies]
+serde = { workspace = true }
+serde_json = { workspace = true }
+
 declavatar = { workspace = true }
 
 [build-dependencies]
