+++ declavac/src/state.rs
@@ -1,10 +1,10 @@
-use crate::{
+use crate::serialization::Jsoned;
+
+use declavatar::{
     avatar_v2::{data::avatar::Avatar, Transformer},
     decl_v2::{compile_declaration, Arguments, DeclarationFormat},
     log::{Log, SerializedLog},
 };
-
-use serde::Serialize;
 use serde_json::Error as SerdeJsonError;
 
 #[derive(Debug, Clone)]
@@ -72,37 +72,5 @@ impl CompiledState {
         self.logs
             .get(index)
             .map(|a| a.json().expect("should be serialized"))
-    }
-}
-
-#[derive(Debug, Clone)]
-struct Jsoned<T> {
-    data: T,
-    serialized: Option<String>,
-}
-
-impl<T: Serialize> Jsoned<T> {
-    pub fn new(data: T) -> Result<Jsoned<T>, SerdeJsonError> {
-        let serialized = Some(serde_json::to_string(&data)?);
-        Ok(Jsoned { data, serialized })
-    }
-
-    pub fn data(&self) -> &T {
-        &self.data
-    }
-
-    pub fn json(&self) -> Option<&str> {
-        self.serialized.as_deref()
-    }
-
-    pub fn update(&mut self, f: impl FnOnce(&mut T)) -> Result<(), SerdeJsonError> {
-        f(&mut self.data);
-        self.serialize()?;
-        Ok(())
-    }
-
-    pub fn serialize(&mut self) -> Result<(), SerdeJsonError> {
-        self.serialized = Some(serde_json::to_string(&self.data)?);
-        Ok(())
     }
 }
